// next/src/components/footer/emec-badge/api/types.ts
export type EMecItemDTO = {
  id: number;
  link?: string | null;
  qrcodeUrl: string | null;
  qrcodeAlt?: string | null;
};

export type EMecResponseDTO = {
  data: EMecItemDTO[];
  meta: {
    pagination?: {
      page: number;
      pageSize: number;
      pageCount: number;
      total: number;
    };
  };
};